Split location utilities from model
